# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI Process Flow Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Company {
  id: string;
  name: string;
  ediId: string;
  type: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CompanyInput {
  name: string;
  ediId: string;
  type: string;
  address?: string;
  contactEmail?: string;
  contactPhone?: string;
  isActive?: boolean;
}

export interface CompanyUpdate {
  name?: string;
  ediId?: string;
  type?: string;
  address?: string;
  contactEmail?: string;
  contactPhone?: string;
  isActive?: boolean;
}

export interface PartnerEndpoint {
  id: string;
  companyId: string;
  /** @nullable */
  companyName?: string | null;
  name: string;
  url: string;
  authType: string;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  bearerToken?: string | null;
  /** @nullable */
  customHeaders?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface PartnerEndpointInput {
  companyId: string;
  name: string;
  url: string;
  authType: string;
  apiKey?: string;
  bearerToken?: string;
  customHeaders?: string;
  isActive?: boolean;
}

export interface PartnerEndpointUpdate {
  name?: string;
  url?: string;
  authType?: string;
  apiKey?: string;
  bearerToken?: string;
  customHeaders?: string;
  isActive?: boolean;
}

export interface EdiDocument {
  id: string;
  documentType: string;
  direction: string;
  status: string;
  senderId: string;
  /** @nullable */
  senderName?: string | null;
  receiverId: string;
  /** @nullable */
  receiverName?: string | null;
  controlNumber: string;
  /** @nullable */
  referenceNumber?: string | null;
  /** @nullable */
  poNumber?: string | null;
  /** @nullable */
  shipDate?: string | null;
  /** @nullable */
  deliveryDate?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  /** @nullable */
  lineItems?: string | null;
  /** @nullable */
  paymentTerms?: string | null;
  /** @nullable */
  shippingDetails?: string | null;
  /** @nullable */
  x12Content?: string | null;
  /** @nullable */
  notes?: string | null;
  retryCount: number;
  /** @nullable */
  lastResponseCode?: number | null;
  /** @nullable */
  lastResponseBody?: string | null;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  transactionId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface EdiDocumentInput {
  documentType: string;
  direction: string;
  senderId: string;
  receiverId: string;
  referenceNumber?: string;
  poNumber?: string;
  shipDate?: string;
  deliveryDate?: string;
  totalAmount?: number;
  lineItems?: string;
  paymentTerms?: string;
  shippingDetails?: string;
  notes?: string;
  transactionId?: string;
  status?: string;
}

export interface EdiDocumentUpdate {
  documentType?: string;
  status?: string;
  referenceNumber?: string;
  poNumber?: string;
  shipDate?: string;
  deliveryDate?: string;
  totalAmount?: number;
  lineItems?: string;
  paymentTerms?: string;
  shippingDetails?: string;
  notes?: string;
  x12Content?: string;
}

export interface SendResult {
  success: boolean;
  message: string;
  /** @nullable */
  responseCode?: number | null;
  /** @nullable */
  responseBody?: string | null;
  /** @nullable */
  sentAt?: string | null;
}

export interface X12Preview {
  content: string;
  documentType: string;
}

export interface InboundPayload {
  x12Content: string;
}

export interface InboundReceipt {
  success: boolean;
  messageId: string;
  /** @nullable */
  documentType?: string | null;
  /** @nullable */
  sender?: string | null;
  /** @nullable */
  receiver?: string | null;
  errors?: string[];
}

export interface InboundMessage {
  id: string;
  /** @nullable */
  documentType?: string | null;
  /** @nullable */
  senderId?: string | null;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  receiverId?: string | null;
  /** @nullable */
  receiverName?: string | null;
  rawPayload: string;
  /** @nullable */
  parsedData?: string | null;
  status: string;
  /** @nullable */
  validationErrors?: string | null;
  /** @nullable */
  controlNumber?: string | null;
  createdAt: string;
  /** @nullable */
  processedAt?: string | null;
}

export interface Transaction {
  id: string;
  referenceNumber: string;
  status: string;
  initiatorId: string;
  /** @nullable */
  initiatorName?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  totalValue?: number | null;
  documents?: EdiDocument[];
  createdAt: string;
  updatedAt: string;
}

export interface DashboardSummary {
  totalDocuments: number;
  outboundCount: number;
  inboundCount: number;
  deliveredCount: number;
  failedCount: number;
  draftCount: number;
  pendingCount: number;
  activeTransactions: number;
  companiesCount: number;
}

export interface FlowStep {
  step: string;
  documentType: string;
  sender: string;
  receiver: string;
  status: string;
  sequenceOrder: number;
  /** @nullable */
  documentId?: string | null;
  /** @nullable */
  timestamp?: string | null;
}

export interface ActivityItem {
  id: string;
  type: string;
  documentType: string;
  direction: string;
  status: string;
  senderName: string;
  receiverName: string;
  /** @nullable */
  referenceNumber?: string | null;
  /** @nullable */
  totalAmount?: number | null;
  createdAt: string;
}

export interface DocumentStat {
  documentType: string;
  total: number;
  delivered: number;
  failed: number;
  draft: number;
  sent: number;
}

export interface AuditLog {
  id: string;
  action: string;
  entityType: string;
  entityId: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export type ListEdiDocumentsParams = {
status?: string;
direction?: string;
documentType?: string;
companyId?: string;
};

export type ListInboundMessagesParams = {
status?: string;
documentType?: string;
};

export type ListTransactionsParams = {
status?: string;
companyId?: string;
};

export type ListAuditLogsParams = {
entityType?: string;
entityId?: string;
};