# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SERMACROPS EDI Process Flow Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partner-endpoints:
    get:
      operationId: listPartnerEndpoints
      tags: [partnerEndpoints]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerEndpoint"
    post:
      operationId: createPartnerEndpoint
      tags: [partnerEndpoints]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerEndpointInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerEndpoint"
  /partner-endpoints/{id}:
    get:
      operationId: getPartnerEndpoint
      tags: [partnerEndpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerEndpoint"
    patch:
      operationId: updatePartnerEndpoint
      tags: [partnerEndpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerEndpointUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerEndpoint"
    delete:
      operationId: deletePartnerEndpoint
      tags: [partnerEndpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    PartnerEndpoint:
      type: object
      required: [id, companyId, name, url, authType, isActive, createdAt]
      properties:
        id:
          type: string
        companyId:
          type: string
        companyName:
          type: ["string", "null"]
        name:
          type: string
        url:
          type: string
        authType:
          type: string
        apiKey:
          type: ["string", "null"]
        bearerToken:
          type: ["string", "null"]
        customHeaders:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    PartnerEndpointInput:
      type: object
      required: [companyId, name, url, authType]
      properties:
        companyId:
          type: string
        name:
          type: string
        url:
          type: string
        authType:
          type: string
        apiKey:
          type: string
        bearerToken:
          type: string
        customHeaders:
          type: string
        isActive:
          type: boolean
    PartnerEndpointUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        authType:
          type: string
        apiKey:
          type: string
        bearerToken:
          type: string
        customHeaders:
          type: string
        isActive:
          type: boolean
    EdiDocument:
      type: object
      required: [id, documentType, direction, status, senderId, receiverId, controlNumber, retryCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        documentType:
          type: string
        direction:
          type: string
        status:
          type: string
        senderId:
          type: string
        senderName:
          type: ["string", "null"]
        receiverId:
          type: string
        receiverName:
          type: ["string", "null"]
        controlNumber:
          type: string
        referenceNumber:
          type: ["string", "null"]
        poNumber:
          type: ["string", "null"]
        shipDate:
          type: ["string", "null"]
        deliveryDate:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        lineItems:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        shippingDetails:
          type: ["string", "null"]
        x12Content:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        retryCount:
          type: integer
        lastResponseCode:
          type: ["integer", "null"]
        lastResponseBody:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        transactionId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Transaction:
      type: object
      required: [id, referenceNumber, status, initiatorId, createdAt, updatedAt]
      properties:
        id:
          type: string
        referenceNumber:
          type: string
        status:
          type: string
        initiatorId:
          type: string
        initiatorName:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        totalValue:
          type: ["number", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/EdiDocument"
        createdAt:
          type: string
        updatedAt:
          type: string
...