# lib/api-zod/src/generated/types/partnerEndpoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI Process Flow Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerEndpoint {
  id: string;
  companyId: string;
  /** @nullable */
  companyName?: string | null;
  name: string;
  url: string;
  authType: string;
  /** @nullable */
  apiKey?: string | null;
  /** @nullable */
  bearerToken?: string | null;
  /** @nullable */
  customHeaders?: string | null;
  isActive: boolean;
  createdAt: string;
}