# lib/api-zod/src/generated/types/partnerEndpointInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SERMACROPS EDI Process Flow Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface PartnerEndpointInput {
  companyId: string;
  name: string;
  url: string;
  authType: string;
  apiKey?: string;
  bearerToken?: string;
  customHeaders?: string;
  isActive?: boolean;
}